Updates the script for 16.04.3 execution.
Adds a few checks to avoid double cloning the repos if rerun multiple times.
Fixes a bug where we check for file status not directory status for the xrdp/xorgxrdp repos.
Adds the post script to be run the the logon user account to allow the Unity panels to load.